backend/requirements.txt
alembic==1.13.3
bcrypt==4.2.0
blinker==1.8.2
click==8.1.7
distlib==0.3.8
filelock==3.16.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.1
platformdirs==4.3.2
pluggy==1.5.0
PyJWT==2.9.0
pymssql==2.3.1
pytest==8.3.3
python-dotenv==1.0.1
SQLAlchemy==2.0.35
typing_extensions==4.12.2
virtualenv==20.26.4
Werkzeug==3.0.4
